+++ kube/examples/pod_reflector.rs
@@ -5,8 +5,6 @@ use kube::{
     runtime::Reflector,
     Client,
 };
-use std::time::Duration;
-use tokio::time::delay_for;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -17,34 +15,18 @@ async fn main() -> anyhow::Result<()> {
 
     let pods: Api<Pod> = Api::namespaced(client, &namespace);
     let lp = ListParams::default().timeout(10); // short watch timeout in this example
-    let rf = Reflector::new(pods, lp).init().await?;
+    let rf = Reflector::new(pods).params(lp);
 
-    // Can read initial state now:
-    rf.state().await?.into_iter().for_each(|pod| {
-        let name = Meta::name(&pod);
-        let phase = pod.status.unwrap().phase.unwrap();
-        let containers = pod
-            .spec
-            .unwrap()
-            .containers
-            .into_iter()
-            .map(|c| c.name)
-            .collect::<Vec<_>>();
-        info!("Found initial pod {} ({}) with {:?}", name, phase, containers);
-    });
-
-    let cloned = rf.clone();
+    let rf2 = rf.clone(); // read from a clone in a task
     tokio::spawn(async move {
         loop {
-            if let Err(e) = cloned.poll().await {
-                warn!("Poll error: {:?}", e);
-            }
+            // Periodically read our state
+            tokio::time::delay_for(std::time::Duration::from_secs(5)).await;
+            let pods: Vec<_> = rf2.state().await.unwrap().iter().map(Meta::name).collect();
+            info!("Current pods: {:?}", pods);
         }
     });
 
-    loop {
-        delay_for(Duration::from_secs(5)).await;
-        let pods: Vec<_> = rf.state().await?.iter().map(Meta::name).collect();
-        info!("Current pods: {:?}", pods);
-    }
+    rf.run().await?; // run reflector and listen for signals
+    Ok(())
 }