+++ kube/src/api/object.rs
@@ -20,6 +20,11 @@ where
     Modified(K),
     /// Resource was deleted
     Deleted(K),
+    /// Resource bookmark
+    ///
+    /// From [Watch bookmarks](https://kubernetes.io/docs/reference/using-api/api-concepts/#watch-bookmarks)
+    /// NB: This became Beta first in Kubernetes 1.16
+    Bookmark(K),
     /// There was some kind of error
     Error(ErrorResponse),
 }
@@ -33,6 +38,7 @@ where
             WatchEvent::Added(_) => write!(f, "Added event"),
             WatchEvent::Modified(_) => write!(f, "Modified event"),
             WatchEvent::Deleted(_) => write!(f, "Deleted event"),
+            WatchEvent::Bookmark(_) => write!(f, "Bookmark event"),
             WatchEvent::Error(e) => write!(f, "Error event: {:?}", e),
         }
     }