+++ kube/src/runtime/reflector.rs
@@ -2,109 +2,165 @@ use crate::{
     api::{Api, ListParams, Meta, WatchEvent},
     Error, Result,
 };
-use futures::{lock::Mutex, StreamExt, TryStreamExt};
+use futures::{future::FutureExt, lock::Mutex, pin_mut, select, TryStreamExt};
 use serde::de::DeserializeOwned;
-use tokio::time::delay_for;
+use tokio::{
+    signal::{self, ctrl_c},
+    time::delay_for,
+};
 
 use std::{collections::BTreeMap, sync::Arc, time::Duration};
 
 /// A reflection of state for a Kubernetes ['Api'] resource
 ///
-/// This watches and caches a `Resource<K>` by:
-/// - seeding the cache from a large initial list call
-/// - keeping track of initial, and subsequent resourceVersions
-/// - recovering when resourceVersions get desynced
+/// This builds on top of the ['Informer'] by tracking the events received,
+/// via ['Informer::poll']. This object will in fact use .poll() continuously,
+/// and use the results to maintain an up to date state map.
 ///
-/// It exposes it's internal state readably through a getter.
+/// It is prone to the same desync problems as an informer, but it will self-heal,
+/// as best as possible - though this means that you might occasionally see a full
+/// reset (boot equivalent) when network issues are encountered.
+/// During a reset, the state is cleared before it is rebuilt.
+///
+/// The internal state is exposed readably through a getter.
 #[derive(Clone)]
 pub struct Reflector<K>
 where
-    K: Clone + DeserializeOwned + Send + Meta,
+    K: Clone + DeserializeOwned + Meta,
 {
     state: Arc<Mutex<State<K>>>,
-    api: Api<K>,
     params: ListParams,
+    api: Api<K>,
 }
 
 impl<K> Reflector<K>
 where
-    K: Clone + DeserializeOwned + Meta + Send,
+    K: Clone + DeserializeOwned + Meta,
 {
-    /// Create a reflector on an api resource with a set of parameters
-    pub fn new(api: Api<K>, lp: ListParams) -> Self {
+    /// Create a reflector on an api resource
+    pub fn new(api: Api<K>) -> Self {
         Reflector {
             api,
-            params: lp,
+            params: ListParams::default(),
             state: Default::default(),
         }
     }
 
-    /// Initializes with a full list of data from a large initial LIST call
-    pub async fn init(self) -> Result<Self> {
-        info!("Starting Reflector for {}", self.api.resource.kind);
-        self.reset().await?;
-        Ok(self)
+    /// Modify the default watch parameters for the underlying watch
+    pub fn params(mut self, lp: ListParams) -> Self {
+        self.params = lp;
+        self
     }
 
-    /// Run a single watch poll
-    ///
-    /// If this returns an error, it tries a full refresh.
-    /// This is meant to be run continually in a thread/task. Spawn one.
-    pub async fn poll(&self) -> Result<()> {
-        trace!("Watching {}", self.api.resource.kind);
-        if let Err(e) = self.single_watch().await {
-            warn!("Poll error on {}: {}: {:?}", self.api.resource.kind, e, e);
-            // If desynched due to mismatching resourceVersion, retry in a bit
-            let dur = Duration::from_secs(10);
-            delay_for(dur).await;
-            self.reset().await?; // propagate error if this failed..
-        }
+    /// Start the reflectors self-driving polling
+    pub async fn run(self) -> Result<()> {
+        self.reset().await?;
+        loop {
+            // local development needs listening for ctrl_c
+            let ctrlc_fut = ctrl_c().fuse();
+            // kubernetes apps need to listen for SIGTERM (30s warning)
+            use signal::unix::{signal, SignalKind}; // TODO: conditional compile
+            let mut sigterm = signal(SignalKind::terminate()).unwrap();
+            let sigterm_fut = sigterm.recv().fuse();
 
+            // and reflector needs to poll continuously
+            let poll_fut = self.poll().fuse();
+
+            // Then pin then futures to the stack, and wait for any of them
+            pin_mut!(ctrlc_fut, sigterm_fut, poll_fut);
+            select! {
+                ctrlc = ctrlc_fut => {
+                    warn!("Intercepted ctrl_c signal");
+                    return Ok(());
+                },
+                sigterm = sigterm_fut => {
+                    warn!("Intercepted SIGTERM");
+                    return Ok(());
+                }
+                poll = poll_fut => {
+                    // Error handle if not ok, otherwise, we do another iteration
+                    if let Err(e) = poll {
+                        warn!("Poll error on {}: {}: {:?}", self.api.resource.kind, e, e);
+                        // If desynched due to mismatching resourceVersion, retry in a bit
+                        let dur = Duration::from_secs(10);
+                        delay_for(dur).await;
+                        self.reset().await?; // propagate error if this failed..
+                    }
+                }
+            }
+        }
+    }
+
+    /// A single poll call to modify the internal state
+    async fn poll(&self) -> Result<()> {
+        let kind = &self.api.resource.kind;
+        let resource_version = self.state.lock().await.version.clone();
+        trace!("Polling {} from resourceVersion={}", kind, resource_version);
+        let stream = self.api.watch(&self.params, &resource_version).await?;
+        pin_mut!(stream);
+
+        // For every event, modify our state
+        while let Some(ev) = stream.try_next().await? {
+            let mut state = self.state.lock().await;
+            // Informer-like version tracking:
+            match &ev {
+                WatchEvent::Added(o)
+                | WatchEvent::Modified(o)
+                | WatchEvent::Deleted(o)
+                | WatchEvent::Bookmark(o) => {
+                    // always store the last seen resourceVersion
+                    if let Some(nv) = Meta::resource_ver(o) {
+                        trace!("Updating reflector version for {} to {}", kind, nv);
+                        state.version = nv.clone();
+                    }
+                }
+                _ => {}
+            }
+
+            let data = &mut state.data;
+            // Core Reflector logic
+            match ev {
+                WatchEvent::Added(o) => {
+                    debug!("Adding {} to {}", Meta::name(&o), kind);
+                    data.entry(ObjectId::key_for(&o)).or_insert_with(|| o.clone());
+                }
+                WatchEvent::Modified(o) => {
+                    debug!("Modifying {} in {}", Meta::name(&o), kind);
+                    data.entry(ObjectId::key_for(&o)).and_modify(|e| *e = o.clone());
+                }
+                WatchEvent::Deleted(o) => {
+                    debug!("Removing {} from {}", Meta::name(&o), kind);
+                    data.remove(&ObjectId::key_for(&o));
+                }
+                WatchEvent::Bookmark(o) => {
+                    debug!("Bookmarking {} from {}", Meta::name(&o), kind);
+                }
+                WatchEvent::Error(e) => {
+                    warn!("Failed to watch {}: {:?}", kind, e);
+                    return Err(Error::Api(e));
+                }
+            }
+        }
         Ok(())
     }
 
-    /// Read data for users of the reflector
-    ///
-    /// This is instant if you are reading and writing from the same context.
-    pub async fn state(&self) -> Result<Vec<K>> {
-        let state = self.state.lock().await;
-        Ok(state.data.values().cloned().collect::<Vec<K>>())
-    }
-
-    /// Read a single entry by name
-    ///
-    /// Will read in the configured namsepace, or globally on non-namespaced reflectors.
-    /// If you are using a non-namespaced resources with name clashes,
-    /// Try [`Reflector::get_within`] instead.
-    pub fn get(&self, name: &str) -> Result<Option<K>> {
-        let id = ObjectId {
-            name: name.into(),
-            namespace: self.api.resource.namespace.clone(),
-        };
-
-        futures::executor::block_on(async { Ok(self.state.lock().await.data.get(&id).map(Clone::clone)) })
-    }
-
-    /// Read a single entry by name within a specific namespace
-    ///
-    /// This is a more specific version of [`Reflector::get`].
-    /// This is only useful if your reflector is configured to poll across namsepaces.
-    pub fn get_within(&self, name: &str, ns: &str) -> Result<Option<K>> {
-        let id = ObjectId {
-            name: name.into(),
-            namespace: Some(ns.into()),
-        };
-        futures::executor::block_on(async { Ok(self.state.lock().await.data.get(&id).map(Clone::clone)) })
-    }
-
-    /// Reset the state with a full LIST call
+    /// Reset the state of the underlying informer and clear the cache
     pub async fn reset(&self) -> Result<()> {
-        trace!("Refreshing {}", self.api.resource.kind);
+        trace!("Resetting {}", self.api.resource.kind);
+        // Simplified for k8s >= 1.16
+        //*self.state.lock().await = Default::default();
+        //self.informer.reset().await
+
+        // For now:
         let (data, version) = self.get_full_resource_entries().await?;
         *self.state.lock().await = State { data, version };
         Ok(())
     }
 
+    /// Legacy helper for kubernetes < 1.16
+    ///
+    /// Needed to do an initial list operation because of https://github.com/clux/kube-rs/issues/219
+    /// Soon, this goes away as we drop support for k8s < 1.16
     async fn get_full_resource_entries(&self) -> Result<(Cache<K>, String)> {
         let res = self.api.list(&self.params).await?;
         let version = res.metadata.resource_version.unwrap_or_default();
@@ -128,56 +184,44 @@ where
         Ok((data, version))
     }
 
-    // Watch helper
-    async fn single_watch(&self) -> Result<()> {
-        let rg = &self.api.resource;
-        let oldver = self.state.lock().await.version.clone();
-        let mut events = self.api.watch(&self.params, &oldver).await?.boxed();
+    /// Read data for users of the reflector
+    ///
+    /// This is instant if you are reading and writing from the same context.
+    pub async fn state(&self) -> Result<Vec<K>> {
+        let state = self.state.lock().await;
+        Ok(state.data.values().cloned().collect::<Vec<K>>())
+    }
 
-        // Follow docs conventions and store the last resourceVersion
-        // https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes
-        while let Some(ev) = events.try_next().await? {
-            // Update in place:
-            let mut state = self.state.lock().await;
-            match ev {
-                WatchEvent::Added(o) => {
-                    debug!("Adding {} to {}", Meta::name(&o), rg.kind);
-                    state
-                        .data
-                        .entry(ObjectId::key_for(&o))
-                        .or_insert_with(|| o.clone());
-                    if let Some(v) = Meta::resource_ver(&o) {
-                        state.version = v.to_string();
-                    }
-                }
-                WatchEvent::Modified(o) => {
-                    debug!("Modifying {} in {}", Meta::name(&o), rg.kind);
-                    state
-                        .data
-                        .entry(ObjectId::key_for(&o))
-                        .and_modify(|e| *e = o.clone());
-                    if let Some(v) = Meta::resource_ver(&o) {
-                        state.version = v.to_string();
-                    }
-                }
-                WatchEvent::Deleted(o) => {
-                    debug!("Removing {} from {}", Meta::name(&o), rg.kind);
-                    state.data.remove(&ObjectId::key_for(&o));
-                    if let Some(v) = Meta::resource_ver(&o) {
-                        state.version = v.to_string();
-                    }
-                }
-                WatchEvent::Error(e) => {
-                    warn!("Failed to watch {}: {:?}", rg.kind, e);
-                    return Err(Error::Api(e));
-                }
-            }
-        }
-        Ok(())
+    /// Read a single entry by name
+    ///
+    /// Will read in the configured namsepace, or globally on non-namespaced reflectors.
+    /// If you are using a non-namespaced resources with name clashes,
+    /// Try [`Reflector::get_within`] instead.
+    pub async fn get(&self, name: &str) -> Result<Option<K>> {
+        let id = ObjectId {
+            name: name.into(),
+            namespace: self.api.resource.namespace.clone(),
+        };
+
+        Ok(self.state.lock().await.data.get(&id).map(Clone::clone))
+    }
+
+    /// Read a single entry by name within a specific namespace
+    ///
+    /// This is a more specific version of [`Reflector::get`].
+    /// This is only useful if your reflector is configured to poll across namsepaces.
+    pub async fn get_within(&self, name: &str, ns: &str) -> Result<Option<K>> {
+        let id = ObjectId {
+            name: name.into(),
+            namespace: Some(ns.into()),
+        };
+        Ok(self.state.lock().await.data.get(&id).map(Clone::clone))
     }
 }
 
 /// ObjectId represents an object by name and namespace (if any)
+///
+/// This is an internal subset of ['k8s_openapi::api::core::v1::ObjectReference']
 #[derive(Ord, PartialOrd, Hash, Eq, PartialEq, Clone)]
 struct ObjectId {
     name: String,
@@ -202,11 +246,9 @@ impl ObjectId {
     }
 }
 
-
-/// Internal representation for Reflector
-type Cache<K> = BTreeMap<ObjectId, K>;
-
 /// Internal shared state of Reflector
+///
+/// Can remove this in k8s >= 1.16 once this uses Informer
 struct State<K> {
     data: Cache<K>,
     version: String,
@@ -220,3 +262,5 @@ impl<K> Default for State<K> {
         }
     }
 }
+/// Internal representation for Reflector
+type Cache<K> = BTreeMap<ObjectId, K>;